# back/ManualMovementsManager/ManualMovementsManager/src/ManualMovementsManager.Application/Commands/ManualMovements/AddManualMovement/AddManualMovementResponse.cs
using ManualMovementsManager.Domain.Entities;
using ManualMovementsManager.Domain.Entities.Response;
using ManualMovementsManager.Application.DTOs;

namespace ManualMovementsManager.Application.Commands.ManualMovements.AddManualMovement
{
    public class AddManualMovementResponse : BaseResponse<ManualMovementDto>
    {
    }
} 