# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ManualMovementsManager API",
...
    "termsOfService": "https://yourcompany.com/terms",
    "contact": {
      "name": "Development Team",
      "url": "https://yourcompany.com",
      "email": "dev@yourcompany.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0.0"
  },
  "paths": {
...
    "/api/v1/ManualMovement": {
      "get": {
        "tags": [
          "ManualMovements"
        ],
        "summary": "Listar movimentos manuais",
        "description": "Retorna uma lista paginada de movimentos manuais com filtros opcionais. Suporta paginação e busca por descrição, período e valor.",
        "operationId": "GetManualMovements",
        "parameters": [
          {
            "name": "Description",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinValue",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MaxValue",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "FieldName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Define o idioma preferido (ex: en-US, pt-BR, es)",
            "schema": {
              "type": "string",
              "default": "pt-BR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de movimentos manuais retornada com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              }
            }
          },
          "206": {
            "description": "Lista parcial de movimentos manuais (paginação)",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              }
            }
          },
          "204": {
            "description": "Nenhum movimento manual encontrado"
          },
          "400": {
            "description": "Parâmetros de consulta inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetManualMovementResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ManualMovements"
        ],
        "summary": "Criar novo movimento manual",
        "description": "Cria um novo movimento manual com validação de produtos e produtos COSIF.",
        "operationId": "CreateManualMovement",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Define o idioma preferido (ex: en-US, pt-BR, es)",
            "schema": {
              "type": "string",
              "default": "pt-BR"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddManualMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddManualMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddManualMovementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movimento manual criado com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "422": {
            "description": "Entidade não processável",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddManualMovementResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddManualMovementRequest": {
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "cosifCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "movementDate": {
            "type": "string",
            "format": "date-time"
          },
          "userCode": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AddManualMovementResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ManualMovement"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataStatus": {
        "enum": [
          "created",
          "active",
          "inactive",
          "deleted"
        ],
        "type": "string"
      },
...
      "Exception": {
        "type": "object",
        "properties": {
          "targetSite": {
            "$ref": "#/components/schemas/MethodBase"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "data": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "innerException": {
            "$ref": "#/components/schemas/Exception"
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "hResult": {
            "type": "integer",
            "format": "int32"
          },
          "stackTrace": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetManualMovementResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualMovement"
            },
            "nullable": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ManualMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "launchNumber": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "cosifCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "movementDate": {
            "type": "string",
            "format": "date-time"
          },
          "userCode": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "productCosif": {
            "$ref": "#/components/schemas/ProductCosif"
          }
        },
        "additionalProperties": false
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "productCosifs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCosif"
            },
            "nullable": true
          },
          "manualMovements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualMovement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductCosif": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "cosifCode": {
            "type": "string",
            "nullable": true
          },
          "classificationCode": {
            "type": "string",
            "nullable": true
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "manualMovements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualMovement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}